+++ consensus/tendermint/core/timeout.go
@@ -2,12 +2,12 @@ package core
 
 import (
 	"context"
-	"github.com/clearmatics/autonity/log"
 	"math/big"
 	"sync"
 	"time"
 
 	"github.com/clearmatics/autonity/common"
+	"github.com/clearmatics/autonity/log"
 )
 
 const (
